+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## Unreleased - YYYY-MM-DD
 ### Added
 - Config file with default cache time
+- Added a search method to search the Store API
+- Added a cache method to set a number of minutes to cache for, overriding the default one in config
 
 ### Deprecated
 ### Fixed